Moved WidgetTestCase to django.test.testcases.
It'll be used in other test modules.
